// src/api/models/product-search-result.ts
export default class ProductSearchResult {
    alt_images: Array<any>;
    api_type: string;
    brand: string;
    category: string;
    colors: Array<string>;
    description: string;
    id: string;
    image: string;
    md5: string;
    merchant: string;
    price: number;
    title: string;
    url: string;
}